refactor(qbittorrent-e2e): extract bencode helpers
