+++ src/console/ci/qbittorrent/mod.rs
@@ -1,2 +1,3 @@
+pub mod bencode;
 pub mod qbittorrent_client;
 pub mod runner;
